[SIG-4719] hide footer when in amstedam.app (#2256)
* hide footer when in amstedam.app

* changed imports

* implement PR feedback

* remove unncessecary import

* changed Footer name
